Split snippets into files
